ENH: isct_test_ants improvement
Former-commit-id: 0a47e8d8f08e6886170dc837d1898dad9cfce4a7 [formerly 08c65e40cdf6f5b510a5be6db9e62971a847449b [formerly ec2d7f2a2d834f0d1a7be325cbac2980531d37a9]]
Former-commit-id: 0e6488e590ffd88754fd46d3d435cb4e2c3359e6
Former-commit-id: 91103dc6c89e0e82e763367fb1caf0cfc7bffefb
